+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (2021-08-20)
 - added aws-ts-vpc-ecs-autoscaling-lt 
+- added aws-ts-existingvpc-ecs-autoscaling-lt
 
 ## (2021-08-19)
 - added azure-ts-keyvault